+++ app/events/components/RandomFanart.tsx
@@ -6,7 +6,6 @@ import { FaDice } from 'react-icons/fa';
 import { IoSettingsSharp } from 'react-icons/io5';
 
 import Fanart from '@/app/events/components/Fanart';
-import useLocalStorage from '@/app/events/hooks/useLocalStorage';
 import { useRandomFanart } from '@/app/events/service/client/useEventService';
 import Button from '@/components/Button';
 import Checkbox from '@/components/Checkbox';
@@ -16,6 +15,7 @@ import Popover, {
   PopoverHeader,
   PopoverTrigger,
 } from '@/components/Popover';
+import useLocalStorage from '@/hooks/useLocalStorage';
 import type { CheckBoxType } from '@/types';
 
 const initCheckboxValues = {
